+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 3.0.5 [2024-09-06] - Genesis service fix
+### Changed
+- Update BPM version
+
+### Fixed
+- Execution crash when vtl error on genesis service (kraftwerkExecutionContext null)
+
 ## 3.0.4 [2024-08-22] - BPM integration
 ### Changed
 - Metadata parsing shared with Genesis using BPM (Basic Parser of Metadata)